Arranging into devices and devices_lib
