--- ch22-core/src/address_spaces.rs
@@ -1,6 +1,0 @@
-pub mod address_map;
-pub mod addressable_device;
-pub mod io_space;
-pub mod paged_rom;
-pub mod ram;
-pub mod rom;
