+++ ch22-core/src/devices.rs
@@ -1,0 +1,8 @@
+pub mod io_space;
+pub mod js_io_device;
+pub mod js_timer_device;
+pub mod paged_rom;
+pub mod ram;
+pub mod rom;
+pub mod rom_select;
+pub mod static_device;
