+++ ch22-core/src/devices/paged_rom.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 
 use crate::word::Word;
 
-use super::addressable_device::*;
+use crate::devices_lib::addressable_device::AddressableDevice;
 
 const ROM_SIZE: usize = 0x4000;
