+++ ch22-core/src/devices_lib.rs
@@ -1,0 +1,6 @@
+pub mod address_map;
+pub mod addressable_device;
+pub mod io_device;
+pub mod io_device_list;
+pub mod timer_device;
+pub mod timer_device_list;
