+++ ch22-core/src/devices_lib/addressable_device.rs
