+++ ch22-core/src/devices_lib/io_device.rs
