+++ ch22-core/src/devices_lib/timer_device.rs
