--- ch22-core/src/io_devices.rs
@@ -1,4 +1,0 @@
-pub mod io_device;
-pub mod js_io_device;
-pub mod rom_select;
-pub mod static_device;
