+++ ch22-core/src/system.rs
@@ -1,15 +1,15 @@
 use js_sys::Function;
 use wasm_bindgen::prelude::*;
 
-use crate::address_spaces::address_map::*;
-use crate::address_spaces::io_space::*;
 use crate::cpu::*;
 use crate::cycle_manager::*;
+use crate::devices::js_io_device::JsIODevice;
+use crate::devices::js_timer_device::*;
+use crate::devices::static_device::StaticDevice;
+use crate::devices_lib::address_map::*;
+use crate::devices_lib::io_device_list::IODeviceID;
+use crate::devices_lib::timer_device_list::TimerDeviceID;
 use crate::interrupt_type::InterruptType;
-use crate::io_devices::js_io_device::JsIODevice;
-use crate::io_devices::static_device::StaticDevice;
-use crate::timer_devices::js_timer_device::*;
-use crate::timer_devices::timer_device_list::TimerDeviceID;
 use crate::utils;
 
 #[wasm_bindgen]
