--- ch22-core/src/timer_devices.rs
@@ -1,3 +1,0 @@
-pub mod js_timer_device;
-pub mod timer_device;
-pub mod timer_device_list;
